Builds out the structure of the app
- Simple model with just a schedule field which is a string
- Sets up the getF1Schedule function which makes the http request
- Sets up a simple decoder which returns a race name
- Sets up a simple view which just displays the string "Here is your schedule"

This compiles however there is lots more to do.

What I want to be returned by the decoded is a list of string date pairs which are the race name and the date it is on.

The view needs to be built out so you can actually make the request and see the result.

Once it is possible to make the request for the hardcoded year this can be extended to be any chosen year.

Maybe this can also be extended to show last and next race, not just the whole schedule.