Updates view to display a list of races
First step was to create a hard coded list of races in the model. This allowed me to set up the view to display a list.

Then I could dynamically generate the list by decoding the json from the api.

Set up a decoder to get the list of races and then a decoder which resolved one race from a list.

Set up a button in the view with an onClick function. This is the function that retrieves the json.

The last step was to convert a race from a string to a record. This means it can have a name and a date. I could then set up a decoder to retrieve the date string and display it.